+++ CHANGES.rst
@@ -7,6 +7,10 @@
 
 - Added support for Python 3.6.
 - Dropped support for Python 3.3.
+- Use the adapter namespace from ``zope.pagetemplate`` if it's
+  available, instead of the backwards compatibility shim in
+  ``zope.app.pagetemplate``. See `issue 3
+  <https://github.com/zopefoundation/z3c.pt/issues/3>`_.
 
 
 3.0 (2016-09-02)